ai-agent-core/pyproject.toml
[project]
name = "ai-agent-core"
version = "0.1.0"
description = "AI Agent Core System - Phase 1 MVP"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.3",
    "openai>=1.10.0",
    "anthropic>=0.40.0",
    "httpx>=0.26.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
]
